requirements.txt
django>=6.0,<6.1
djangorestframework>=3.16,<3.17
django-cors-headers>=4.9,<5.0
django-filter>=25.2,<26.0
gunicorn>=23.0,<24.0
whitenoise>=6.8,<7.0
dj-database-url>=2.2,<3.0
psycopg-binary>=3.2,<3.3